+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 Release Notes
 ====
 
+# 08-28-2019
+<a href="https://www.nuget.org/packages/dotnext.aspnetcore.cluster/0.4.0">DotNext.AspNetCore.Cluster 0.4.0</a><br/>
+<a href="https://www.nuget.org/packages/dotnext.net.cluster/0.4.0">DotNext.Net.Cluster 0.4.0</a>
+* Heartbeat timeout can be tuned through configuration
+* Optimized Raft state machine
+
 # 08-27-2019
 <a href="https://www.nuget.org/packages/dotnext.aspnetcore.cluster/0.3.5">DotNext.AspNetCore.Cluster 0.3.5</a><br/>
 * Docker support